+++ components/search/src/lib.rs
@@ -72,7 +72,7 @@ fn add_section_to_index(index: &mut Index, section: &Section, library: &Library)
 
     for key in &section.pages {
         let page = library.get_page_by_key(*key);
-        if !page.meta.in_search_index || page.meta.draft {
+        if !page.meta.in_search_index {
             continue;
         }