+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Update livereload
 - Add `hard_link_static` config option to hard link things in the static directory instead of copying
 - Draft pages are not longer rendered
+- Add warning for old style internal links since they would still function
 
 ## 0.8.0 (2019-06-22)